requirements.txt
streamlit
transformers
streamlit-audiorecorder 
